Solve implicit narrowing converison

+++ device/tt_device/protocol/jtag_protocol.cpp
@@ -18,11 +18,11 @@ JtagProtocol::JtagProtocol(std::unique_ptr<JtagDevice> jtag_device, uint8_t jlin
 JtagProtocol::~JtagProtocol() = default;
 
 void JtagProtocol::write_to_device(const void* mem_ptr, tt_xy_pair core, uint64_t addr, size_t size, NocId noc_id) {
-    jtag_device_->write(mmio_id_, mem_ptr, core.x, core.y, addr, size, static_cast<uint64_t>(noc_id));
+    jtag_device_->write(mmio_id_, mem_ptr, core.x, core.y, addr, size, static_cast<uint8_t>(noc_id));
 }
 
 void JtagProtocol::read_from_device(void* mem_ptr, tt_xy_pair core, uint64_t addr, size_t size, NocId noc_id) {
-    jtag_device_->read(mmio_id_, mem_ptr, core.x, core.y, addr, size, static_cast<uint64_t>(noc_id));
+    jtag_device_->read(mmio_id_, mem_ptr, core.x, core.y, addr, size, static_cast<uint8_t>(noc_id));
 }
 
 bool JtagProtocol::write_to_core_range(const void*, tt_xy_pair, tt_xy_pair, uint64_t, uint32_t, NocId) { return false; }
